Communicate with oracles by stdin/stdout.
This changes the oracle handling to use pipes to talk the
oracle.  The arguments are written line by line into stdin and
stdout is interpreted as one term.

In addition of the arguments the first line written to the
oracle is '(',  the second line is the rule name that invoces the
oracle, and the last line is ')'.  This allows us to use the same
oracle script for multiple rules.  Furthermore, we might later
want to have oracles that maintain persistant state.
In this case the () help to distinguise different invocations.