Implement experimental :match-conclusion for rules
When defining a rule, one can use :match-conclusion instead of
:conclusion.  If this is used, the conclusion must be given in
the step.  To achieve this, define-rule with a
`:match-conclusion T` adds an additional argument `Quote T` to
the start of the argument list, and sets the conclusion to
`Proof T`.  The `step` command detects rules annotated with
`:match-conclusion` and hands the conclusion to the rule as
the first argument.

Hence, this feature doesn't change functional semantic of rules,
it is just syntactic sugar for `step`.