Add support for `eo::dt_constructors` and `eo::dt_selectors` to inspect datatypes and datatype constructors (#87)
This adds a way `eo::dt_constructors` to extract the definition of
datatypes (their corresponding list of constructors) and
`eo::dt_selectors` to extract the definition of datatype constructors
(their corresponding list of selectors). It adds an ordinary type
`eo::List` that is part of the assumed background signature for this
purpose.

Also fixes the parser for `declare-datatype` (the singular form of
datatypes, which had not been tested).