Do not return non-ground types with eo::typeof (#124)
`eo::typeof` is an exceptional case in that it can return a term with
free parameters. It is the only such operator that may do so. This
behavior is fairly unintuitive and moreover allows non-first-class terms
e.g. `Quote` to appear in terms.

We instead make it so that `eo::typeof` does not evaluate if the type is
non-ground.