Update type system with annotated parameters (#141)
This changes type checking such that the free parameters of the types of
parameters are also bound when that parameter is instantiated.

The motivation for this change:
- It makes the relationship between quote arrow and ordinary arrow more
intuitive, i.e. quoting an argument does not impact its behavior.
- Ensures the types of parameters are taken into account instead of
ignored.
- Allows for more powerful proof rules, e.g. the array extensionality
rule added in a regression on this PR.

To implement this, we introduce a new distinguished operator `eo::param`
or "annotated parameter" that "wraps" a parameter with its type. The
match algorithm now treats this operator as a special case, matching the
second argument with the type of the term that is matched. This operator
is used internally but is not available in the parser.

Fixes several regressions where the more pedantic check catches an
error.