Change eo::is_eq, add eo::eq and eo::is_ok (#140)
This changes the semantics of `eo::is_eq` such that it returns "false"
for two terms that are both stuck. This also impacts the semantics of
`eo::requires`, whose semantics depends on `eo::is_eq`.

It introduces `eo::eq` and `eo::is_ok`, which are the more primitive
operators that `eo::is_eq` can be described in terms of.