More consistent definition of eo::nil (#151)
Currently the definition of `eo::nil` is very ad-hoc, as it can accept
an arbitrary number of arguments.

This PR updates `eo::nil` so that it *always* requires exactly 2
arguments, the list operator and the return type.

The user manual is updated with a simpler description, which does not
rely on invoking type checking to compute the nil terminator, instead
this can be understood as program evaluation.

---------

Co-authored-by: Hans-Jörg <hansjoergschurr@users.noreply.github.com>