Add an optimization to list ops, fix corner case of list_nth (#155)
Based on the forthcoming spec, `list_nth` should not evaluate if the
argument is not an f-list (even if it is a list up to the index n). The
code was over-optimized to evaluate successfully in this case.

This also adds important optimizations to erase/erase_all/setof/concat
to avoid unecessary expression construction, which in my initial tests
makes e.g. chain resolution 15% faster.