Drop support for eo::match (#164)
We drop support for eo::match since we do not have a way to type check
such terms.

Match terms were always just syntax sugar for auxiliary functions, so
this should be used as an alternative.

CPC on the current main branch of cvc5 has eliminated all occurrences of
eo::match.