Better support for explicit conclusions (#153)
This adds support for `:conclusion-explicit` and drops support for
`:conclusion-given` as well as the distinguished variable
`eo::conclusion`.

This also makes our checking for assumptions / step-pop matching more
pedantic.

For example, this proof checked previously by treating an ordinary
argument as an assumption, this now fails due to an explicit check that
proof rules with `:assumption` are the only ones that take assumptions.
```
(declare-rule temp ((A Bool))
  :assumption A
  :conclusion true
)
(step @p2 true :rule temp :args (true))