Forbid evaluation in types of parameters (#173)
This forbids examples such as the following:
```
(declare-parameterized-const foo1 
  ((T Type) (U Type) (F Bool) (c (eo::ite F T U)))
  Type)

(program foo2 ((T Type) (U Type) (F Bool) (t T))  
  :signature ((eo::ite F T U)) Type
  (
  ((foo2 t) Type)
  )
)
```
The intuitive reason is that argument types is analogous to patterns in
program evaluation, where evaluation is forbidden.

Notice that these restrictions forbid the combination of quote and
evaluatable types. The restriction does *not* forbid the use of
evaluation in any function argument type. In particular, it is important
to note that the following example is still permitted which is
equivalent to the first example:
```
(declare-parameterized-const foo1b 
  ((T Type) (U Type) (F Bool)) 
  (-> (eo::ite F T U) Type))  ; still allowed
```

On the other hand, this example is disallowed and has no equivalent
after this PR:
```
(declare-const Set (-> Type Type))
(declare-parameterized-const foo3
  ((T Type) (U Type) (F Bool) (c (eo::ite F T U)) (w Type)) 
  (Set w))
```