Proof rules as programs (#183)
This changes the internal implementation of ethos so that proof rules
are (references to) programs, not constants.

This makes the implementation more uniform and will lead to several
forthcoming simplifications.

Furthermore updates a regression which was based off of a BV RARE rule
with implicit evaluation in premises, using a similar fix as applied to
the one in official Cpc.eo.

---------

Co-authored-by: Hans-Jörg <hansjoergschurr@users.noreply.github.com>